renamed targetsource package to discovery
