+++ internal/controller/discovery/loaders/all/all.go
@@ -1,6 +1,6 @@
 package all
 
 import (
-	_ "github.com/gnmic/operator/internal/controller/targetsource/loaders/http_pull"
+	_ "github.com/gnmic/operator/internal/controller/discovery/loaders/http_pull"
 	// _ "github.com/gnmic/operator/internal/controller/targetsource/loaders/http_push"
 )
