+++ internal/controller/discovery/loaders/http_pull/loader_test.go
