+++ internal/controller/discovery/mapper.go
@@ -1,4 +1,4 @@
-package targetsource
+package discovery
 
 // This file makes diff between existing and new targets
 // file decides which targets to create/update/delete
